requirements.txt
kafka-python==2.0.2
hdbcli==2.20.23
pyyaml==6.0.1
requests==2.31.0
